pyproject.toml
[tool.poetry]
name = "vocode"
version = "0.1.111"
description = "The all-in-one voice SDK"
authors = ["Ajay Raj <ajay@vocode.dev>"]
license = "MIT License"
readme = "README.md"
homepage = "https://github.com/vocodedev/vocode-python"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
pydub = "^0.25.1"
nltk = "^3.8.1"
openai = "^0.27.8"
sounddevice = "^0.4.6"
azure-cognitiveservices-speech = "^1.27.0"
websockets = "^11.0.2"
requests = "^2.28.2"
uvicorn = "^0.21.1"
fastapi = "^0.100.1"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
six = "^1.16.0"
opentelemetry-sdk = "^1.17.0"
janus = "^1.0.0"
scipy = "^1.10.1"
anthropic = "^0.2.9"
aiofiles = "^23.2.1"

elevenlabs = {version = "^0.2.6", optional = true}
google-cloud-texttospeech = {version = "^2.14.1", optional = true}
gtts = {version = "^2.3.1", optional = true}

google-cloud-speech = {version = "^2.19.0", optional = true}

redis = {version = "^4.5.4", optional = true}
twilio = {version = "^8.1.0", optional = true}
vonage = {version = "^3.5.1", optional = true}
nylas = {version = "^5.14.0", optional = true}
speechrecognition = "^3.10.0"
aiohttp = "^3.8.4"
langchain = "^0.0.198"
google-cloud-aiplatform = {version = "^1.26.0", optional = true}
miniaudio = "^1.59"
boto3 = "^1.28.28"


[tool.poetry.group.lint.dependencies]
black = "^23.1.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
aioresponses = "^0.7.4"
pre-commit = "^3.3.3"


[tool.poetry.group.typing.dependencies]
mypy = "^1.3.0"
pydub-stubs = "^0.25.1.0"
types-requests = "^2.30.0.0"
types-regex = "^2023.5.5.0"
types-redis = "^4.5.5.2"

[tool.mypy]
ignore_missing_imports = "True"
exclude = ["venv", ".venv"]
explicit_package_bases = "True"


[tool.poetry.extras]
synthesizers = ["gtts", "google-cloud-texttospeech", "elevenlabs"]
transcribers = ["google-cloud-speech"]
telephony = ["twilio", "redis", "vonage"]
agents = ["google-cloud-aiplatform"]
actions = ["nylas"]
all = ["gtts", "google-cloud-texttospeech", "elevenlabs", "google-cloud-speech", "google-cloud-aiplatform", "twilio", "redis", "nylas", "vonage"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
